+++ lib/src/toml_localizations.dart
@@ -6,43 +6,77 @@ import 'package:flutter/services.dart' show rootBundle;
 class TomlLocalizations {
   final _tomlParser = TomlParser();
 
-  /// map of translations per languageCode
+  /// map of translations per language/country code
   final Map<String, Map> _translationMap = {};
 
-  /// path to Yaml translation asset
+  /// path to translation assets
   final String assetPath;
 
-  /// language code of current locale, set in [load] method
-  String _languageCode;
+  /// a hash key of language / country code used for [_translationsMap]
+  String _codeKey;
 
   /// initialize with asset path to Toml files and a list of supported language codes
   TomlLocalizations(this.assetPath);
 
-  /// first time we call load, we read the csv file and initialize translations
-  /// next time we just return this
-  /// called by [TomlLocalizationsDelegate]
+  Future<String> loadAsset(path) async {
+    try {
+      return await rootBundle.loadString(path);
+    } catch (e) {}
+    return null;
+  }
+
+  /// load and cache a toml file per language / country code
   Future<TomlLocalizations> load(Locale locale) async {
-    this._languageCode = locale.languageCode;
-    if (_translationMap.containsKey(_languageCode)) {
+    final languageCode = locale.languageCode;
+    final countryCode = locale.countryCode;
+
+    assert(languageCode != null);
+    assert(languageCode.isNotEmpty);
+
+    if (countryCode != null && countryCode.isNotEmpty) {
+      _codeKey = '$languageCode-$countryCode';
+    } else {
+      _codeKey = '$languageCode';
+    }
+    debugPrint(_codeKey);
+
+    // in cache
+    if (_translationMap.containsKey(_codeKey)) {
       return this;
     }
-    final String path = '$assetPath/$_languageCode.toml';
-    final String text = await rootBundle.loadString(path);
-    final Map<String, dynamic> toml = _tomlParser.parse(text).value;
-    _translationMap[_languageCode] = toml;
+
+    // try to load with with _codeKey
+    // could be a combination of language / country code
+    debugPrint('load $_codeKey');
+    final text = await loadAsset('$assetPath/$_codeKey.toml');
+    if (text != null) {
+      _translationMap[_codeKey] = _tomlParser.parse(text).value;
+      return this;
+    }
+
+    // if it was a combined key, try to load with only language code
+    if (_codeKey != languageCode) {
+      _codeKey = languageCode;
+      debugPrint('load $_codeKey');
+      final text = await loadAsset('$assetPath/$_codeKey.toml');
+      // asset file should always exist for a supportedLanguageCode
+      assert(text != null);
+      _translationMap[_codeKey] = _tomlParser.parse(text).value;
+    }
+
+    assert(false, 'translation file not found for code \'$_codeKey\'');
+
     return this;
   }
 
   /// get translation given a key
   String string(String key) {
-    // find translation map given current locale
-    final bool containsLocale = _translationMap.containsKey(_languageCode);
-    assert(containsLocale, 'Missing localization for code: $_languageCode');
-    final Map translations = _translationMap[_languageCode];
-    // find translated string given translation key
-    final bool containsKey = translations.containsKey(key);
+    final containsLocale = _translationMap.containsKey(_codeKey);
+    assert(containsLocale, 'Missing localization for code: $_codeKey');
+    final translations = _translationMap[_codeKey];
+    final containsKey = translations.containsKey(key);
     assert(containsKey, 'Missing localization for translation key: $key');
-    final String translatedValue = translations[key];
+    final translatedValue = translations[key];
     return translatedValue;
   }