+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [0.3.0] - 30 november 2020
+- support country codes
 - remove supportedLanguageCodes and simplify constructor
 
 ## [0.2.2] - 19 november 2020